Potential fix for code scanning alert no. 11: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be achieved by casting one of the operands to int64_t before performing the multiplication. Specifically, we can cast trie.dataLength to int64_t to ensure the multiplication is done in a larger type.

 - Change the multiplication trie.dataLength * (dataIs32 ? 4 : 2) to (int64_t)trie.dataLength * (dataIs32 ? 4 : 2).
 - This change should be made on line 68 in the file libicu4c-64-2/common/utrie_swap.cpp.

+++ libicu4c-64-2/common/utrie_swap.cpp
@@ -65,7 +65,7 @@ utrie_swap(const UDataSwapper *ds,
     }
 
     dataIs32=(UBool)((trie.options&UTRIE_OPTIONS_DATA_IS_32_BIT)!=0);
-    size=sizeof(UTrieHeader)+trie.indexLength*2+trie.dataLength*(dataIs32?4:2);
+    size=sizeof(UTrieHeader)+trie.indexLength*2+(int64_t)trie.dataLength*(dataIs32?4:2);
 
     if(length>=0) {
         UTrieHeader *outTrie;
